implemented ListPackageFiles
