+++ src/dpkg_list_package_files.rs
@@ -1,4 +1,6 @@
 use crate::dpkg_options::DpkgOptions;
+use std::io::Error;
+use std::process::Command;
 
 pub struct ListPackageFiles {
     package_name_pattern: String,
@@ -14,7 +16,21 @@ impl ListPackageFiles {
         ListPackageFiles { package_name_pattern, options: Some(options) }
     }
 
-    fn run() {
-        todo!()
+    pub fn exec(&self) -> Result<Vec<String>, Error> {
+        let mut command = Command::new("dpkg-query");
+
+        // applying dpkg Options
+        match &self.options {
+            Some(options) => command.args(options.build()),
+            _ => &mut command // TOOD: Why does that have to be here?
+        };
+
+        command.args(["--listfiles", &self.package_name_pattern]);
+
+        println!("Executing {:?}", &command);
+        match command.output() {
+            Ok(data) => { Ok(String::from_utf8_lossy(&data.stdout).split("\n").filter(move |x| { !x.is_empty() }).map(|s| s.to_string()).collect()) }
+            Err(e) => Err(e)
+        }
     }
 }
